requirements.txt
django
djangorestframework
djangorestframework-simplejwt
pymongo>=3.12.0
dnspython>=2.1.0
mysqlclient
mongoengine==0.28.2
transformers==4.44.2
torch==2.4.0
pillow==10.4.0
nltk==3.9.1
requests==2.32.3